ItemSet library started
I have the major structures for the item sets (serializing them).
I wrote the start of a class for writing out the item sets to the disk:
It can find where League is installed (via registry), has the default
paths to global and champion specific sets. Can get the full path of the
sets and set the league of legends install directory (if we wanted the
user to browse for it).