Removed extra property and added future css
So there was this extra bit of css that wasn't need and then after doing
some research I could have used @media to not have to write the
javascript. This tag can make window size specific css allow you to
support mobile. It's probably cleaner to go through and add some more
css, but work so hard on the JS.